+++ Robust.Server/Console/Commands/MapCommands.cs
@@ -1,9 +1,6 @@
-using System.Globalization;
 using System.Linq;
-using System.Text;
 using Robust.Server.GameObjects;
 using Robust.Server.Maps;
-using Robust.Server.Player;
 using Robust.Shared.Console;
 using Robust.Shared.ContentPack;
 using Robust.Shared.GameObjects;
@@ -14,62 +11,6 @@ using Robust.Shared.Maths;
 
 namespace Robust.Server.Console.Commands
 {
-    sealed class AddMapCommand : LocalizedCommands
-    {
-        [Dependency] private readonly IMapManager _map = default!;
-
-        public override string Command => "addmap";
-
-        public override void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            if (args.Length < 1)
-                return;
-
-            var mapId = new MapId(int.Parse(args[0]));
-
-            if (!_map.MapExists(mapId))
-            {
-                _map.CreateMap(mapId);
-                if (args.Length >= 2 && args[1] == "false")
-                {
-                    _map.AddUninitializedMap(mapId);
-                }
-
-                shell.WriteLine($"Map with ID {mapId} created.");
-                return;
-            }
-
-            shell.WriteError($"Map with ID {mapId} already exists!");
-        }
-    }
-
-    sealed class RemoveMapCommand : LocalizedCommands
-    {
-        [Dependency] private readonly IMapManager _map = default!;
-
-        public override string Command => "rmmap";
-
-        public override void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            if (args.Length != 1)
-            {
-                shell.WriteError("Wrong number of args.");
-                return;
-            }
-
-            var mapId = new MapId(int.Parse(args[0]));
-
-            if (!_map.MapExists(mapId))
-            {
-                shell.WriteError($"Map {mapId.Value} does not exist.");
-                return;
-            }
-
-            _map.DeleteMap(mapId);
-            shell.WriteLine($"Map {mapId.Value} was removed.");
-        }
-    }
-
     public sealed class SaveGridCommand : LocalizedCommands
     {
         [Dependency] private readonly IEntityManager _ent = default!;
@@ -381,167 +322,6 @@ namespace Robust.Server.Console.Commands
                 shell.WriteLine(Loc.GetString("cmd-loadmap-success", ("mapId", mapId), ("path", args[1])));
             else
                 shell.WriteLine(Loc.GetString("cmd-loadmap-error", ("path", args[1])));
-        }
-    }
-
-    sealed class LocationCommand : LocalizedCommands
-    {
-        [Dependency] private readonly IEntityManager _ent = default!;
-
-        public override string Command => "loc";
-
-        public override void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            var player = shell.Player as IPlayerSession;
-            var pt = player?.AttachedEntityTransform;
-            if (pt == null)
-                return;
-
-            var pos = pt.Coordinates;
-
-            shell.WriteLine(
-                $"MapID:{pos.GetMapId(_ent)} GridUid:{pos.GetGridUid(_ent)} X:{pos.X:N2} Y:{pos.Y:N2}");
-        }
-    }
-
-    sealed class TpGridCommand : LocalizedCommands
-    {
-        [Dependency] private readonly IEntityManager _ent = default!;
-        [Dependency] private readonly IMapManager _map = default!;
-
-        public override string Command => "tpgrid";
-
-        public override void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            if (args.Length is < 3 or > 4)
-            {
-                shell.WriteError($"Usage: {Help}");
-                return;
-            }
-
-            var gridId = EntityUid.Parse(args[0]);
-            var xPos = float.Parse(args[1], CultureInfo.InvariantCulture);
-            var yPos = float.Parse(args[2], CultureInfo.InvariantCulture);
-
-            if (!_map.TryGetGrid(gridId, out var grid))
-            {
-                shell.WriteError($"No grid found with id {args[0]}");
-                return;
-            }
-
-            var gridXform = _ent.GetComponent<TransformComponent>(grid.Owner);
-            var mapId = args.Length == 4 ? new MapId(int.Parse(args[3])) : gridXform.MapID;
-
-            gridXform.Coordinates = new EntityCoordinates(_map.GetMapEntityId(mapId), new Vector2(xPos, yPos));
-
-            shell.WriteLine("Grid was teleported.");
-        }
-    }
-
-    sealed class RemoveGridCommand : LocalizedCommands
-    {
-        [Dependency] private readonly IMapManager _map = default!;
-
-        public override string Command => "rmgrid";
-
-        public override void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            if (args.Length != 1)
-            {
-                shell.WriteError("Wrong number of args.");
-                return;
-            }
-
-            var gridId = EntityUid.Parse(args[0]);
-
-            if (!_map.GridExists(gridId))
-            {
-                shell.WriteError($"Grid {gridId} does not exist.");
-                return;
-            }
-
-            _map.DeleteGrid(gridId);
-            shell.WriteLine($"Grid {gridId} was removed.");
-        }
-    }
-
-    internal sealed class RunMapInitCommand : LocalizedCommands
-    {
-        [Dependency] private readonly IMapManager _map = default!;
-
-        public override string Command => "mapinit";
-
-        public override void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            if (args.Length != 1)
-            {
-                shell.WriteError("Wrong number of args.");
-                return;
-            }
-
-            var arg = args[0];
-            var mapId = new MapId(int.Parse(arg, CultureInfo.InvariantCulture));
-
-            if (!_map.MapExists(mapId))
-            {
-                shell.WriteError("Map does not exist!");
-                return;
-            }
-
-            if (_map.IsMapInitialized(mapId))
-            {
-                shell.WriteError("Map is already initialized!");
-                return;
-            }
-
-            _map.DoMapInitialize(mapId);
-        }
-    }
-
-    internal sealed class ListMapsCommand : LocalizedCommands
-    {
-        [Dependency] private readonly IMapManager _map = default!;
-
-        public override string Command => "lsmap";
-
-        public override void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            var msg = new StringBuilder();
-
-            foreach (var mapId in _map.GetAllMapIds().OrderBy(id => id.Value))
-            {
-                msg.AppendFormat("{0}: init: {1}, paused: {2}, ent: {3}, grids: {4}\n",
-                    mapId, _map.IsMapInitialized(mapId),
-                    _map.IsMapPaused(mapId),
-                    _map.GetMapEntityId(mapId),
-                    string.Join(",", _map.GetAllMapGrids(mapId).Select(grid => grid.Owner)));
-            }
-
-            shell.WriteLine(msg.ToString());
-        }
-    }
-
-    internal sealed class ListGridsCommand : LocalizedCommands
-    {
-        [Dependency] private readonly IEntityManager _ent = default!;
-        [Dependency] private readonly IMapManager _map = default!;
-
-        public override string Command => "lsgrid";
-        public override void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            var msg = new StringBuilder();
-            var xformQuery = _ent.GetEntityQuery<TransformComponent>();
-
-            foreach (var grid in _map.GetAllGrids().OrderBy(grid => grid.Owner))
-            {
-                var xform = xformQuery.GetComponent(grid.Owner);
-                var worldPos = xform.WorldPosition;
-
-                msg.AppendFormat("{0}: map: {1}, ent: {2}, pos: {3:0.0},{4:0.0} \n",
-                    grid.Owner, xform.MapID, grid.Owner, worldPos.X, worldPos.Y);
-            }
-
-            shell.WriteLine(msg.ToString());
         }
     }
 }
