+++ Robust.Shared/Console/IConsoleCommand.cs
@@ -36,6 +36,11 @@ namespace Robust.Shared.Console
         string Help { get; }
 
         /// <summary>
+        /// If true, this command will be unavailable to clients while they are connected to a server. Has no effect on servers.
+        /// </summary>
+        bool RequireServerOrSingleplayer => false;
+
+        /// <summary>
         /// Executes the client command.
         /// </summary>
         /// <param name="shell">The console that executed this command.</param>
