+++ Robust.Shared/Console/LocalizedCommands.cs
@@ -1,4 +1,4 @@
-﻿using System.Threading;
+using System.Threading;
 using System.Threading.Tasks;
 using Robust.Shared.IoC;
 using Robust.Shared.Localization;
@@ -17,6 +17,9 @@ public abstract class LocalizedCommands : IConsoleCommand
 
     /// <inheritdoc />
     public virtual string Help => LocalizationManager.TryGetString($"cmd-{Command}-help", out var val) ? val : "";
+
+    /// <inheritdoc />
+    public virtual bool RequireServerOrSingleplayer => false;
 
     /// <inheritdoc />
     public abstract void Execute(IConsoleShell shell, string argStr, string[] args);
